Add files via upload
Value At Risk (VAR) with SQL
The Value at Risk (VaR) is a financial metric that aims to summarize the total risk of a portfolio of financial assets in a single number. More specifically, VaR is a statistical technique used to measure the amount of potential loss that could happen in an investment portfolio over a specified period. It is defined as the maximum amount excepted to be lost over a given time horizon at a confidence level.
Example: Suppose the VaR at a 95% confidence level and one-month horizon time is $1 million. Then we are sure at 95% that we cannot lose more than $1 million in one month.
VaR can be calculate using different method: Parametric, Monte Carlo, and Historical. You can fin below  to   use SQL to calculate VaR with the historical method for a portfolio with a single position.